requirements.txt
pydantic==1.10.14
streamlit-pydantic
psycopg2-binary==2.9.9